+++ lib/tabs/mobile/mobile_tab_view.dart
@@ -76,9 +76,16 @@ class MobileTabView extends StatelessWidget {
     return GetBuilder(
       init: router,
       builder: (_) => AnimatedBackground(
-        scale:
-            router.mode == AppMode.projects && router.project != null ? 0 : 0.3,
+        scale: _backgroundScale(),
       ),
     );
+  }
+
+  // METHODS ===================================================================
+
+  double _backgroundScale() {
+    if (router.mode == AppMode.projects && router.project != null) return 0;
+    if (router.mode == AppMode.blogs && router.blog != null) return 0;
+    return 0.3;
   }
 }